Update libraries and fix CI
- Upgrades landingi/php-coding-standards, fixing hundreds of PHPStan deprecations
- Moves integrations tests to actual Integration suite, instead of having them in the Functional one
- Adds IntegrationTestCase with a half-second delay between tests, to limit the request rate to external services
- Adds Codecov github action on both PR and pushes to master
