+++ Makefile
@@ -1,7 +1,11 @@
 ci:
 	vendor/bin/phpstan analyse -c phpstan.neon --memory-limit=256M
 	vendor/bin/ecs check --config vendor/landingi/php-coding-standards/ecs.php
-	vendor/bin/phpunit --testsuite=all --coverage-clover=build/coverage.xml
+	vendor/bin/phpunit --testsuite=unit,functional --coverage-clover=var/coverage/coverage.xml
+ci-all:
+	vendor/bin/phpstan analyse -c phpstan.neon --memory-limit=256M
+	vendor/bin/ecs check --config vendor/landingi/php-coding-standards/ecs.php
+	vendor/bin/phpunit --testsuite=all --coverage-clover=var/coverage/coverage.xml
 fix:
 	vendor/bin/ecs check --fix --config vendor/landingi/php-coding-standards/ecs.php
 test:
@@ -15,7 +19,7 @@ integration:
 coverage:
 	vendor/bin/phpunit --testsuite=all --coverage-text
 coverage-html:
-	vendor/bin/phpunit --testsuite=all --coverage-html=build/coverage/
+	vendor/bin/phpunit --testsuite=all --coverage-html=var/coverage/
 analyse:
 	vendor/bin/phpstan analyse -c phpstan.neon --memory-limit=256M
 	vendor/bin/ecs check --config vendor/landingi/php-coding-standards/ecs.php
