+++ composer.json
@@ -12,7 +12,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
-        "landingi/php-coding-standards": "1.0.4",
+        "landingi/php-coding-standards": "1.0.14",
         "symfony/var-dumper": "^5.3"
     },
     "autoload": {
@@ -25,5 +25,10 @@
             "Landingi\\BookkeepingBundle\\" : "tests"
         }
     },
-    "minimum-stability": "stable"
+    "minimum-stability": "stable",
+    "config": {
+        "allow-plugins": {
+            "phpstan/extension-installer": true
+        }
+    }
 }
