+++ tests/Integration/Wfirma/Contractor/WfirmaContractorBookTest.php
@@ -1,21 +1,21 @@
 <?php
 declare(strict_types=1);
 
-namespace Landingi\BookkeepingBundle\Functional\Wfirma\Contractor;
+namespace Landingi\BookkeepingBundle\Integration\Wfirma\Contractor;
 
 use Generator;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Company;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\ContractorBook;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Person;
+use Landingi\BookkeepingBundle\Integration\IntegrationTestCase;
 use Landingi\BookkeepingBundle\Memory\Contractor\Company\ValueAddedTax\MemoryIdentifierFactory;
 use Landingi\BookkeepingBundle\Wfirma\Client\Credentials\WfirmaCredentials;
 use Landingi\BookkeepingBundle\Wfirma\Client\WfirmaClient;
 use Landingi\BookkeepingBundle\Wfirma\Contractor\Factory\ContractorFactory;
 use Landingi\BookkeepingBundle\Wfirma\Contractor\WfirmaContractorBook;
-use PHPUnit\Framework\TestCase;
 
-final class WfirmaContractorBookTest extends TestCase
+final class WfirmaContractorBookTest extends IntegrationTestCase
 {
     private ContractorBook $book;
