+++ tests/Integration/Wfirma/Invoice/DownloadInvoiceTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Landingi\BookkeepingBundle\Functional\Wfirma\Invoice;
+namespace Landingi\BookkeepingBundle\Integration\Wfirma\Invoice;
 
 use DateTime;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor;
@@ -9,12 +9,14 @@ use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Address\City;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Address\Country;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Address\PostalCode;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Address\Street;
+use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Company;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\ContractorAddress;
+use Landingi\BookkeepingBundle\Bookkeeping\Contractor\ContractorBook;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\ContractorEmail;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\ContractorIdentifier;
 use Landingi\BookkeepingBundle\Bookkeeping\Contractor\ContractorName;
-use Landingi\BookkeepingBundle\Bookkeeping\Contractor\Person;
 use Landingi\BookkeepingBundle\Bookkeeping\Currency;
+use Landingi\BookkeepingBundle\Bookkeeping\Invoice;
 use Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceBook;
 use Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceDescription;
 use Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceFullNumber;
@@ -26,8 +28,10 @@ use Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceItem\ValueAddedTax;
 use Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceSeries;
 use Landingi\BookkeepingBundle\Bookkeeping\Invoice\InvoiceTotalValue;
 use Landingi\BookkeepingBundle\Bookkeeping\Language;
+use Landingi\BookkeepingBundle\Integration\IntegrationTestCase;
 use Landingi\BookkeepingBundle\Memory\Contractor\Company\ValueAddedTax\MemoryIdentifierFactory;
 use Landingi\BookkeepingBundle\Wfirma\Client\Credentials\WfirmaCredentials;
+use Landingi\BookkeepingBundle\Wfirma\Client\Exception\NotFoundException;
 use Landingi\BookkeepingBundle\Wfirma\Client\WfirmaClient;
 use Landingi\BookkeepingBundle\Wfirma\Contractor\Factory\ContractorFactory;
 use Landingi\BookkeepingBundle\Wfirma\Contractor\WfirmaContractorBook;
@@ -37,12 +41,12 @@ use Landingi\BookkeepingBundle\Wfirma\Invoice\WfirmaInvoiceBook;
 use Landingi\BookkeepingBundle\Wfirma\Invoice\WfirmaInvoiceItem;
 use Landingi\BookkeepingBundle\Wfirma\Invoice\WfirmaInvoiceItemCollection;
 use Landingi\BookkeepingBundle\Wfirma\WfirmaInvoice;
-use PHPUnit\Framework\TestCase;
 
-final class WfirmaInvoiceBookTest extends TestCase
+class DownloadInvoiceTest extends IntegrationTestCase
 {
-    private Contractor\ContractorBook $contractorBook;
+    private ContractorBook $contractorBook;
     private InvoiceBook $invoiceBook;
+    private DateTime $today;
 
     public function setUp(): void
     {
@@ -58,60 +62,72 @@ final class WfirmaInvoiceBookTest extends TestCase
         );
         $this->invoiceBook = new WfirmaInvoiceBook($client, new InvoiceFactory(), $factory);
         $this->contractorBook = new WfirmaContractorBook($client, $factory);
+        $this->today = new DateTime();
     }
 
-    public function testInvoiceWorkflow(): void
+    public function testDownloadInvoice(): void
     {
-        $contractor = $this->getContractor();
-
+        // Arrange
+        $contractor = $this->contractorBook->create(
+            new Company(
+                new ContractorIdentifier('123'),
+                new ContractorName('test foo'),
+                new ContractorEmail('test@landingi.com'),
+                new ContractorAddress(
+                    new Street('test 123'),
+                    new PostalCode('11-111'),
+                    new City('test'),
+                    new Country('PL')
+                ),
+                new Company\ValueAddedTax\SimpleIdentifier('6762461659')
+            )
+        );
         $invoice = $this->invoiceBook->create(
             new WfirmaInvoice(
                 new InvoiceIdentifier('123'),
                 new InvoiceSeries(new InvoiceSeries\InvoiceSeriesIdentifier(0)),
-                new InvoiceDescription('test description - bundle invoice'),
+                new InvoiceDescription('testCompanyInPoland'),
                 new InvoiceFullNumber('FV 69/2021'),
                 new InvoiceTotalValue(100),
                 new WfirmaInvoiceItemCollection([
                     new WfirmaInvoiceItem(
                         new Name('foo 1'),
                         new Price((int) (100.55 * 100)),
-                        new WfirmaValueAddedTax('222', new ValueAddedTax(23)),
+                        new WfirmaValueAddedTax('0', new ValueAddedTax(23)),
                         new NumberOfUnits(2)
                     ),
                 ]),
                 $contractor,
                 new Currency('PLN'),
-                new DateTime(),
-                new DateTime(),
-                new DateTime(),
+                $this->today,
+                $this->today,
+                $this->today,
                 new Language('PL')
             )
         );
 
-        self::assertNotEmpty($invoice->getIdentifier()->toString());
+        // Act
+        $invoiceFile = base64_encode($this->invoiceBook->download($invoice->getIdentifier()));
 
-        //test find
-        $invoice = $this->invoiceBook->find($invoice->getIdentifier());
+        //Assert
+        $this->assertNotEmpty($invoiceFile);
+        $this->assertIsString($invoiceFile);
 
-        //test delete
-        $this->invoiceBook->delete($invoice->getIdentifier());
-        $this->contractorBook->delete($contractor->getIdentifier());
+        $this->cleanUp($invoice, $contractor);
     }
 
-    private function getContractor(): Contractor
+    public function testItFailsOnDownloadFile(): void
     {
-        return $this->contractorBook->create(
-            new Person(
-                new ContractorIdentifier('123'),
-                new ContractorName('test foo'),
-                new ContractorEmail('test@landingi.com'),
-                new ContractorAddress(
-                    new Street('test 123'),
-                    new PostalCode('111-111'),
-                    new City('test'),
-                    new Country('PL')
-                )
-            )
-        );
+        // Arrange
+        $this->expectException(NotFoundException::class);
+
+        // Act
+        $this->invoiceBook->download(new InvoiceIdentifier('1'));
+    }
+
+    private function cleanUp(Invoice $invoice, Contractor $contractor): void
+    {
+        $this->invoiceBook->delete($invoice->getIdentifier());
+        $this->contractorBook->delete($contractor->getIdentifier());
     }
 }
